Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/text-orientation-parsing-sideways-right-001-expected.txt


FAIL text-orientation: sideways-right assert_equals: expected "sideways" but got "mixed"
